+++ src/chanjo2/endpoints/report.py
@@ -1,13 +1,30 @@
 import logging
 from os import path
+from typing import Dict, List
 
-from fastapi import APIRouter, Depends
+from fastapi import APIRouter, Depends, Request
 from fastapi.responses import HTMLResponse
 from fastapi.templating import Jinja2Templates
 from sqlmodel import Session
 
+from chanjo2.crud.intervals import get_genes
+from chanjo2.crud.samples import get_samples_coverage_file
+from chanjo2.dbutil import Base as SQLModelBase
 from chanjo2.dbutil import get_session
-from chanjo2.models.pydantic_models import ReportQuery
+from chanjo2.meta.handle_d4 import (
+    get_genes_coverage_completeness,
+    get_gene_interval_coverage_completeness,
+)
+from chanjo2.models.pydantic_models import ReportQuery, CoverageInterval
+from chanjo2.models.sql_models import Exon as SQLExon
+from chanjo2.models.sql_models import Gene as SQLGene
+from chanjo2.models.sql_models import Transcript as SQLTranscript
+
+INTERVAL_TYPE_DB_TYPE: Dict[str, SQLModelBase] = {
+    "genes": SQLGene,
+    "transcripts": SQLTranscript,
+    "exons": SQLExon,
+}
 
 LOG = logging.getLogger("uvicorn.access")
 
@@ -19,8 +36,10 @@ router = APIRouter()
 
 
 @router.post("/report/", response_class=HTMLResponse)
-async def report(query: ReportQuery, db: Session = Depends(get_session)):
-    """Return a coverage report over a list of genes for a list of  samples."""
+async def report(
+        request: Request, query: ReportQuery, db: Session = Depends(get_session)
+):
+    """Return a coverage report over a list of genes for a list of samples."""
 
     samples_d4_files: Tuple[str, D4File] = get_samples_coverage_file(
         db=db, samples=query.samples, case=query.case
@@ -35,6 +54,21 @@ async def report(query: ReportQuery, db: Session = Depends(get_session)):
         limit=None,
     )
 
-    LOG.warning(query.interval_type)
+    interval_type = INTERVAL_TYPE_DB_TYPE[query.interval_type]
 
-    return templates.TemplateResponse("item.html")
+    if interval_type == SQLGene:
+        cov_compl_data: List[CoverageInterval] = get_genes_coverage_completeness(
+            samples_d4_files=samples_d4_files,
+            genes=genes,
+            completeness_threholds=query.completeness_thresholds,
+        )
+    else:
+        cov_compl_data: List[CoverageInterval] = get_gene_interval_coverage_completeness(
+            db=db,
+            samples_d4_files=samples_d4_files,
+            genes=genes,
+            interval_type=SQLTranscript,
+            completeness_threholds=query.completeness_thresholds,
+        )
+
+    return templates.TemplateResponse("item.html", {"request": request, "data": cov_compl_data})